requirements.txt
streamlit>=1.35.0
anthropic>=0.34.2
python-dotenv>=1.0.1
numpy
